% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{as_logical_index}
\alias{as_logical_index}
\title{Assure a full length logical index}
\usage{
as_logical_index(index, n)
}
\arguments{
\item{index}{A vector, typically a numeric row index of the data to
coerce to a logical index.}

\item{n}{Single numeric, the number of rows of the data use as a replicate
return length.}
}
\description{
Suggests a alpha opacity to plot with as a function of the number of
observation.
}
\examples{
## Coerce a numeric index to logical:
as_logical_index(c(1, 4:10, 15), nrow(mtcars))
}
