% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_visualization.r
\name{basis_local_attribution}
\alias{basis_local_attribution}
\title{Extract and format the 1D local attribution basis}
\usage{
basis_local_attribution(attr_df, rownum = nrow(attr_df))
}
\arguments{
\item{attr_df}{Return of a local attribution, such as
treeshap_df.}

\item{rownum}{The rownumber of the primary observation.}
}
\value{
A matrix of the 1D basis
}
\description{
Internal function, Extract and format the 1D local attribution basis from
the provided local explanation's attribution.
}
\examples{
la_df <- mtcars ## Pretend this is a local attribution data.frame
basis_local_attribution(la_df, rownum = 10)
}
