% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheem-package.r
\docType{package}
\name{cheem}
\alias{cheem}
\title{Trees of {cheem}}
\description{
Consider an arbitrary black-box model. Local explanations are a class of
point-measures of the variable importance to the model for one location in
the explanatory variables. \code{Cheem} extracts the local attribution of all
observations and first creates a 2D approximated space comparing data- and
attribution-spaces. After a primary and comparison observation have been
selected the attribution of the primary is used as the 1D basis for a manual
tour. This tour changes the contribution from the variable that deviates the
most from its expected value. By viewing the positions of the  primary and
comparison points, one can interrogate the explanation identify improperly
weighted variable especially in the case of of a misclassified point as
compared with a correctly classified neighbor.
}
\details{
GitHub: \url{https://github.com/nspyrison/cheem}
}
\seealso{
\code{\link[=cheem_ls]{cheem_ls()}} \code{\link[=run_app]{run_app()}}
}
