% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{does_contain_nonnumeric}
\alias{does_contain_nonnumeric}
\title{Check if a vector contains non-numeric character}
\usage{
does_contain_nonnumeric(char)
}
\description{
Returns a logical, whether or not a vector contains any non-numeric
characters. Typically used to test if row names hold non-index information.
}
