% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_attribution_ls.r
\name{format_ls}
\alias{format_ls}
\title{Format the nested local attribution layers}
\usage{
format_ls(
  local_attr_ls,
  x,
  y,
  basis_type = c("pca", "olda"),
  class = NULL,
  keep_model = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{local_attr_ls}{A return from \code{local_attr_ls()}.}

\item{x}{The explanatory variables of the model.}

\item{y}{The target variable of the model.}

\item{basis_type}{The type of basis used to approximate the data and
attribution space from. Defaults to "pca".}

\item{class}{The variable to group points by. Originally the \emph{predicted}
class.}

\item{keep_model}{Whether or not to keep the (sizable) model object.
Defaults to FALSE.}

\item{verbose}{Logical, Whether or not the function should print tictoc time
info.}
}
\value{
A list of formated data frames.
}
\description{
Internal function, formats all layers of all data.frames, consumed in
local_attr_ls().
}
