% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_list.r
\name{format_nested_layers}
\alias{format_nested_layers}
\title{Format the nested local attribution layers}
\usage{
format_nested_layers(
  layer_ls,
  x,
  y,
  basis_type = c("pca", "olda"),
  class = NULL,
  verbose = TRUE
)
}
\arguments{
\item{layer_ls}{A return from \code{nested_local_attr_layers()}.}

\item{x}{The explanatory variables of the model.}

\item{y}{The target variable of the model.}

\item{basis_type}{The type of basis used to approximate the data and
attribution space from. Defaults to "pca".}

\item{class}{The variable to group points by. Originally the \emph{predicted}
class.}

\item{verbose}{Logical, Whether or not the function should print tictoc time
info.}
}
\value{
A list of formated data frames.
}
\description{
Internal function, formats all layers of all data.frames, consumed in
nested_local_attr_layers().
}
