% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{global_view_1layer}
\alias{global_view_1layer}
\title{Create the plot data.frame for the global linked plotly display.}
\usage{
global_view_1layer(
  x,
  y,
  class = NULL,
  basis_type = c("pca", "olda"),
  layer_name = class(x)[1]
)
}
\arguments{
\item{x}{The explanatory variables of the model.}

\item{y}{The target variable of the model.}

\item{class}{The variable to group points by. Originally the \emph{predicted}
class.}

\item{basis_type}{The type of basis used to approximate the data and
attribution space from. Defaults to "pca".}

\item{layer_name}{Character layer name, typically the type of local
attribution used.}
}
\value{
A data.frame, for the global linked plotly display.
}
\description{
Internal function, the plot data.frame of 1 layer, consumed downstream in
cheem_ls.
}
\examples{
sub <- DALEX::apartments[1:200, 1:6]
x <- sub[, 2:5]
y <- sub$m2.price
clas <- sub$district

ret <- global_view_1layer(
  x, y, class = clas,
  basis_type = "pca", layer_name = "data")
str(ret)
}
