% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_attribution_ls.r, R/1_cheem_lists.r
\name{global_view_1sp}
\alias{global_view_1sp}
\title{Create the plot data.frame for the global linked plotly display.}
\usage{
global_view_1sp(
  x,
  y,
  basis_type = c("pca", "olda"),
  class = NULL,
  layer_name = class(x)[1]
)

global_view_1sp(
  x,
  y,
  basis_type = c("pca", "olda"),
  class = NULL,
  layer_name = class(x)[1]
)
}
\arguments{
\item{x}{The explanatory variables of the model.}

\item{y}{The target variable of the model.}

\item{basis_type}{The type of basis used to approximate the data and
attribution space from. Defaults to "pca".}

\item{class}{The variable to group points by. Originally the \emph{predicted}
class.}

\item{layer_name}{Character layer name, typically the type of local
attribution used.}
}
\value{
A data.frame, for the global linked plotly display.

A data.frame, for the global linked plotly display.
}
\description{
Internal function, the plot data.frame of 1 layer, consumed in
local_attr_ls() and format_ls().

Internal function, the plot data.frame of 1 layer, consumed in
local_attr_ls() and format_ls().
}
\examples{
sub <- DALEX::apartments[1:200, 1:6]
x <- sub[, 2:5]
y <- sub$m2.price
clas <- sub$district

ret <- global_view_1sp(x, y, basis_type = "pca",
                       class = clas, layer_name ="SHAP")
str(ret)
sub <- DALEX::apartments[1:200, 1:6]
x <- sub[, 2:5]
y <- sub$m2.price
clas <- sub$district

ret <- global_view_1sp(x, y, basis_type = "pca",
                       class = clas, layer_name ="SHAP")
str(ret)
}
