% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{ifDev}
\alias{ifDev}
\title{Evaluate if development}
\usage{
ifDev(expr)
}
\arguments{
\item{text}{A character string to message() if package version is 9000.}
}
\description{
Evaluate the epression if the 4th chunk of the package version is 9000.
}
