% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{is.discrete}
\alias{is.discrete}
\title{Check if a vector is discrete}
\usage{
is.discrete(x)
}
\description{
Returns a logical, whether or not a vector is a discrete variable.
Typically used on the Y variable of a model.
}
\details{
\subsection{see plyr::is.discrete(). !! not on levels, class only}{
}

\subsection{some RF code may be looking for 5 of fewer unique values.}{
}
}
