% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_treeshap_minimal_port.r
\name{is.model_unified}
\alias{is.model_unified}
\title{Check whether object is a valid model_unified object}
\usage{
is.model_unified(x)
}
\arguments{
\item{x}{an object to check}
}
\value{
boolean
}
\description{
Does not check correctness of representation, only basic checks
}
