% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{is_discrete}
\alias{is_discrete}
\title{Check if a vector is discrete}
\usage{
is_discrete(x)
}
\arguments{
\item{x}{A vector to check the discreteness of.}
}
\value{
Logical, whether or not \code{x} is a discrete variable,
}
\description{
Whether or not a vector is a discrete variable, returns a logical.
Typically used on the Y variable of a model.
}
\examples{
is_discrete(mtcars$mpg) ## Numeric column, too many levels for discretness
is_discrete(mtcars$cyl) ## Numeric column, labeled as discrete, because less than 25 unique values
is_discrete(letters)    ## Characters and factors labeled discrete.
}
\seealso{
Other cheem utility: 
\code{\link{as_logical_index}()},
\code{\link{basis_attr_df}()},
\code{\link{does_contain_nonnumeric}()},
\code{\link{linear_tform}()},
\code{\link{logistic_tform}()},
\code{\link{problem_type}()},
\code{\link{rnorm_from}()}
}
\concept{cheem utility}
