% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{linear_tform}
\alias{linear_tform}
\title{Linear function to help set alpha opacity}
\usage{
linear_tform(n, appox_max_n = 5000L, ceiling = 1, floor = 0.2)
}
\arguments{
\item{n}{Number of observations to plot.}

\item{appox_max_n}{The number of observation to reach floor opacity.}

\item{ceiling}{The highest number returned. Defaults to 1.}

\item{floor}{The lowest number returned. Defaults to 0.2.}
}
\value{
A scalar numeric, suggested value to set alpha opacity.
}
\description{
Suggests a alpha opacity to plot with as a function of the number of
observation.
}
\examples{
## Suggest an opacity to use in plotting:
(my_alpha <- linear_tform(nrow(spinifex::penguins)))

## Visualize
x <- 1:2000
plot(x, sapply(x, linear_tform), col = 'blue')
}
