% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_visualization.r
\name{linked_plotly_func}
\alias{linked_plotly_func}
\title{Linked \code{plotly} display, global view of data and attribution space.}
\usage{
linked_plotly_func(
  layer_ls,
  primary_obs = NULL,
  comparison_obs = NULL,
  height_px = 640L,
  width_px = 640L,
  do_include_maha_qq = FALSE
)
}
\arguments{
\item{layer_ls}{A return from \code{nested_local_attr_layers()}, a list of data frames.}

\item{primary_obs}{The rownumber of the primary observation. Its local
attribution becomes the 1d projection basis, and the point it highlighted
as a dashed line.}

\item{comparison_obs}{The rownumber of the comparison observation. Point
is highlighted as a dotted line.}

\item{height_px}{The height in pixels of the returned \code{plotly} plot.}

\item{width_px}{The width in pixels of the returned \code{plotly} plot.}

\item{do_include_maha_qq}{Logical, whether or not to add the qq plots of the
Mahalanobis distance for the data- and attribution- spaces}
}
\value{
\code{plotly} plot of the global view, first 2 components of the basis of
the data- and attribution- spaces.
}
\description{
Given an attribution layer list, create a linked \code{plotly}of the global data-
and attribution- spaces. Typically consumed directly by shiny app.
}
\examples{
sub <- DALEX::apartments[1:200, 1:6]
X <- sub[, 2:5]
Y <- sub$m2.price
clas <- sub$district

layer_ls <- nested_local_attr_layers(
  x=X, y=Y, n_layers=1, basis_type="pca", class=clas, verbose=T, noisy=T)

linked_plotly_func(layer_ls, primary_obs = 1, comparison_obs = 2)
}
