% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_attribution_ls.r, R/1_cheem_lists.r
\name{local_attr_ls}
\alias{local_attr_ls}
\title{Create the local attribution layer data.frame}
\usage{
local_attr_ls(
  x,
  y,
  xtest = NULL,
  ytest = NULL,
  layer_name = "UNAMED",
  basis_type = c("pca", "olda"),
  class = NULL,
  verbose = TRUE,
  noisy = TRUE
)

local_attr_ls(
  x,
  y,
  xtest = NULL,
  ytest = NULL,
  layer_name = "UNAMED",
  basis_type = c("pca", "olda"),
  class = NULL,
  verbose = TRUE,
  noisy = TRUE
)
}
\arguments{
\item{x}{The explanatory variables of the model.}

\item{y}{The target variable of the model.}

\item{xtest}{Optional, out Of Sample data to find the local attribution of.}

\item{ytest}{Optional, out Of Sample response to measure x with
if provided.}

\item{layer_name}{Character layer name, typically the type of local
attribution used.}

\item{basis_type}{The type of basis used to approximate the data and
attribution space from. Expects "pca" or "olda" (requires \code{clas}).
Defaults to "pca".}

\item{class}{The variable to group points by. Originally the \emph{predicted}
class.}

\item{verbose}{Logical, Whether or not the function should print tictoc time
info. Defaults to TRUE.}

\item{noisy}{Logical, Whether of not the function should play a beeper tone
upon completion. Defaults to TRUE.}
}
\value{
A data.frame, for the full local attribution matrix.

A data.frame, for the full local attribution matrix.
}
\description{
Internal function, the local attribution layer data.frame of 1 layer, consumed in
local_attr_ls().

Internal function, the local attribution layer data.frame of 1 layer, consumed in
local_attr_ls().
}
\examples{
library(cheem)
sub <- DALEX::apartments[1:200, 1:6]
x <- sub[, 2:5]
y <- sub$m2.price
clas <- sub$district

ret <- local_attr_ls(
  x, y, layer_name = "typically 'data'/'<attribution name>'",
  basis_type = "pca", class = clas)
names(ret)
library(cheem)
sub <- DALEX::apartments[1:200, 1:6]
x <- sub[, 2:5]
y <- sub$m2.price
clas <- sub$district

ret <- local_attr_ls(
  x, y, layer_name = "typically 'data'/'<attribution name>'",
  basis_type = "pca", class = clas)
names(ret)
}
