% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_visualization.r
\name{manip_var_of_attr_df}
\alias{manip_var_of_attr_df}
\title{Find the manip var from a given attr_df}
\usage{
manip_var_of_attr_df(attr_df, primary_obs, comparison_obs)
}
\arguments{
\item{attr_df}{A data frame of local explanation attributions,
such as a return from \code{attr_df_treeshap()}.}

\item{rownum}{The rownumber of the primary observation. Defaults to 1.}
}
\value{
A single number of the variable with the largest difference
}
\description{
Find the number of the variable with the largest differnce between the
primary and comparison observations.
}
\examples{
library(cheem)

## Regression:
dat <- amesHousing2018_NorthAmes
X <- dat[, 1:9]
Y <- log(dat$SalePrice)
clas <- dat$SubclassMS

rf_fit <- default_rf(X, Y)
## Long runtime for full datasets:
shap_df <- attr_df_treeshap(rf_fit, X, noisy = FALSE)
manip_var_of_attr_df(shap_df, primary_obs = 1, comparison_obs = 2)
}
\seealso{
Other cheem utility: 
\code{\link{as_logical_index}()},
\code{\link{basis_attr_df}()},
\code{\link{does_contain_nonnumeric}()},
\code{\link{is_discrete}()},
\code{\link{linear_tform}()},
\code{\link{logistic_tform}()},
\code{\link{problem_type}()},
\code{\link{rnorm_from}()}
}
\concept{cheem utility}
