% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_app_plotting_functions.r
\name{manual_tour1d_func}
\alias{manual_tour1d_func}
\title{Cheem tour; 1D manual tour on the selected attribution}
\usage{
manual_tour1d_func(
  layer_ls,
  basis,
  mv_name,
  primary_obs,
  comparison_obs = NULL,
  do_add_pcp_segements = TRUE,
  pcp_shape = c(142, 124),
  angle = 0.1
)
}
\arguments{
\item{layer_ls}{A return from \code{nested_local_attr_layers()}, a list of data frames.}

\item{primary_obs}{The rownumber of the primary observation. Its local
attribution becomes the 1d projection basis, and the point it highlighted
as a dashed line.}

\item{comparison_obs}{The rownumber of the comparison observation. Point
is highlighted as a dotted line.}

\item{height_px}{The height in pixels of the returned \code{plotly} plot.}

\item{width_px}{The width in pixels of the returned \code{plotly} plot.}

\item{do_include_maha_qq}{Logical, whether or not to add the qq plots of the
Mahalanobis distance for the data- and attribution- spaces}
}
\value{
\code{plotly} plot of the global view, first 2 components of the basis of
the data- and attribution- spaces.
}
\description{
Create a linked \code{plotly}of the global data-
and attribution- spaces. Typically consumed directly by shiny app.
}
