% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{model_performance}
\alias{model_performance}
\title{Extract higher level model performance statistics}
\usage{
model_performance(y, pred, label = "label")
}
\arguments{
\item{y}{Observed response, required by ranger models.}

\item{pred}{A (n, 1) vector of the predicted values for each observation.
Typically obtainable with \code{stats::predict()}. Exact syntax and arguments
may change across model types.}

\item{label}{Optionally provide a character label to store reminder
text for the type of model and local explanation used.
Defaults to "label".}
}
\value{
A data.frame of model performance statistics.
}
\description{
Internal function, used downstream in cheem_ls.
}
