% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_list.r
\name{nested_local_attr_layers}
\alias{nested_local_attr_layers}
\title{Format the nested local attribution layers}
\usage{
nested_local_attr_layers(
  x,
  y,
  xtest = NULL,
  ytest = NULL,
  n_layers = 1,
  basis_type = c("pca", "olda"),
  class = NULL,
  verbose = TRUE,
  noisy = TRUE
)
}
\arguments{
\item{x}{The explanatory variables of the model.}

\item{y}{The target variable of the model}

\item{xtest}{Optional, Out Of Sample data to find the local attribution of.}

\item{ytest}{Optional, Out Of Sample response to measure xtext with
if provided.}

\item{n_layers}{The number of local attribution layers deep. Defaults to 1,
the local attribution of just the data. In bulk, I think going more layers is
going to overfit the data, so be prepared to optimize OOS measures.}

\item{basis_type}{The type of basis used to approximate the data and
attribution space from. Defaults to "pca".}

\item{class}{The variable to group points by. Originally the \emph{predicted}
class.}

\item{verbose}{Logical, Whether or not the function should print tictoc time
info. Defaults to TRUE.}

\item{noisy}{Logical, Whether or not the function should play a beepr tone
upon completion. Defaults to TRUE.}
}
\value{
A list of formatted data frames.
}
\description{
Internal function, formats all layers of all data.frames, consumed in
nested_local_attr_layers().
}
\examples{
sub <- DALEX::apartments[1:200, 1:6]
set.seed(303)
.idx_test <- sample(
  1:nrow(sub), size = round(.5 * nrow(sub))) ## 20\% index for testing
X_train <- sub[-.idx_test, 2:5]
Y_train <- sub$m2.price[-.idx_test]
clas <- sub$district[-.idx_test]
X_test  <- sub[.idx_test, ]
Y_test  <- sub$m2.price[.idx_test]

layer_ls <- nested_local_attr_layers(
  x=X_train, y=Y_train, xtest=X_test, ytest=Y_test,
  n_layers=1, basis_type="pca", class=clas, verbose=T, noisy=T)
layer_ls$performance_df
}
