% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_treeshap_minimal_port.r
\name{print.model_unified}
\alias{print.model_unified}
\title{Prints model_unified objects}
\usage{
\method{print}{model_unified}(x, ...)
}
\arguments{
\item{x}{a model_unified object}

\item{...}{other arguments}
}
\value{
Prints a data.frame of the $model of the treeshap unified model.
}
\description{
Prints model_unified objects
}
