% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_treeshap_minimal_port.r
\name{print.treeshap}
\alias{print.treeshap}
\title{Prints treeshap objects}
\usage{
\method{print}{treeshap}(x, ...)
}
\arguments{
\item{x}{a treeshap object}

\item{...}{other arguments}
}
\value{
A data.frame of $shaps of the treeshap object. Also prints
interactions if used.
}
\description{
Prints treeshap objects
}
