% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_visualization.r
\name{radial_cheem_ggtour}
\alias{radial_cheem_ggtour}
\title{Cheem tour; 1D manual tour on the selected attribution}
\usage{
radial_cheem_ggtour(
  cheem_ls,
  basis,
  mv_name,
  primary_obs,
  comparison_obs,
  do_add_pcp_segments = TRUE,
  pcp_shape = c(142, 124),
  angle = 0.2,
  row_index = NULL,
  inc_vars = NULL
)
}
\arguments{
\item{cheem_ls}{A return from \code{cheem_ls()}, a list of data frames.}

\item{basis}{A 1D projection basis, typically a return of
\code{basis_local_attribution()}.}

\item{mv_name}{The character string, \emph{name} of the Manipulation Variable.}

\item{primary_obs}{The rownumber of the primary observation. Its local
attribution becomes the 1d projection basis, and the point it highlighted
as a dashed line.}

\item{comparison_obs}{The rownumber of the comparison observation. Point
is highlighted as a dotted line.}

\item{do_add_pcp_segments}{Logical, whether or not to add parallel coordinate
line segments to the basis display.}

\item{pcp_shape}{The number of the shape character to add. Typically
142 or 124, '|' for \code{plotly} and \code{gganimate} respectively. Defaults to 142,
'|' for \code{plotly}.}

\item{angle}{The step size between interpolated frames, in radians.}

\item{row_index}{Numeric index of selected observations.
Defaults to TRUE; 1:n.}

\item{inc_vars}{A vector of the names of the variables to include in the projection.}
}
\value{
\code{plotly} plot of the global view, first 2 components of the basis of
the data- and attribution- spaces.
}
\description{
Create a linked \code{plotly}of the global data-
and attribution- spaces. Typically consumed directly by shiny app.
}
\examples{
sub <- DALEX::apartments[1:200, 1:6]
X <- sub[, 2:5]
Y <- sub$m2.price
clas <- sub$district

cheem_ls <- cheem_ls(
  x=X, y=Y, basis_type="pca", class=clas, verbose=T, noisy=T)

bas <- basis_local_attribution(cheem_ls$attr_df, rownum = 1)
ggt <- radial_cheem_ggtour(
  cheem_ls, basis=bas, mv_name=colnames(X)[1],
  primary_obs=1, comparison_obs=2)
spinifex::animate_plotly(ggt)
}
