% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{rnorm_observation_of}
\alias{rnorm_observation_of}
\title{Check if a vector contains non-numeric character}
\usage{
rnorm_observation_of(data, n_new_obs = 7, sd_coeff = 1)
}
\description{
Returns a logical, whether or not a vector contains any non-numeric
characters. Typically used to test if row names hold non-index information.
}
