% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_attribution_ls.r, R/1_cheem_lists.r
\name{treeshap_df}
\alias{treeshap_df}
\title{Extract the full SHAP matrix of a randomForest model}
\usage{
treeshap_df(randomForest_model, x, verbose = TRUE)

treeshap_df(randomForest_model, x, verbose = TRUE)
}
\arguments{
\item{randomForest_model}{The return of fitted randomForest::randomForest
model.}

\item{x}{The explanatory data (without response) to extract the local
attributions of.}

\item{data}{Data to extract the local attributions of.}
}
\value{
A dataframe of the local attributions.

A dataframe of the local attributions.
}
\description{
Currently internal function, wants to be generalized.
Extracts a lighter version of the treeshap *.unify of a randomForest.

Currently internal function, wants to be generalized.
Extracts a lighter version of the treeshap *.unify of a randomForest.
}
\examples{
sub <- DALEX::apartments[1:200, 1:5]
x <- sub[, 2:5]
y <- sub$m2.price

## Fit a {randomForest} model, 
# model fit slightly slower fit than {ranger},
# but treeshap iis much faster than {ranger}
.rf <- randomForest::randomForest(y ~ ., data = data.frame(y, x))
system.time(
 df_shap <- treeshap_df(.rf, data = xdat)
)[3]
library(cheem)
sub <- DALEX::apartments[1:200, 1:5]
x <- sub[, 2:5]
y <- sub$m2.price

.rf_fit <- default_rf(x, y)
.shap_df <- treeshap_df(.rf_fit, x = x)
}
