% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{unify_predict}
\alias{unify_predict}
\alias{unify_tree_model}
\title{Unified predictions to a standard format}
\usage{
unify_predict(model, x)

unify_tree_model(model, x)
}
\arguments{
\item{model}{A tree based model supported by \code{treeshap}:
a model from \code{randomForest::randomForest}, \code{ranger::ranger}, \code{gbm::gbm},
\code{xgboost::xgb.train}, or \code{lightgbm::lightgbm}.}

\item{x}{The explanatory data (without response) to extract the local
attributions from.}
}
\value{
A vector of predicted values
}
\description{
A wrapper function for stat::predict(). Condition handling for various
treeshap supported models.

Unifies models supported by treeshap into a standard format.
}
\examples{
library(cheem)

## Regression:
dat  <- amesHousing2018_NorthAmes
X    <- dat[, 1:9]
Y    <- log(dat$SalePrice)
clas <- dat$SubclassMS

rf_fit  <- default_rf(X, Y)
## Applies the correct format to a predict function for the model type.
unify_predict(rf_fit, X)

## Applies the correct treeshap::*.unify for the model type.
unif_mod <- unify_tree_model(rf_fit, X)
str(unif_mod)
}
\seealso{
Other cheem unify: 
\code{\link{is_randomForest}()}
}
\concept{cheem unify}
