% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{unify_tree_model}
\alias{unify_tree_model}
\title{#' Check is the model is supported by treeshap.
#'
#'
#' @param model A model to check the treeshap support of.
is_treeshap_supported <- function(model){
any(is_randomForest(model), is_ranger(model), is_gmb(model),
is_xgboost(model), is_catboost(model), is_lightgbm(model))
}
Unifies models into a standard format}
\usage{
unify_tree_model(model, x)
}
\description{
Unifies models supported by treeshap into a standard format.
}
