% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_visualization.r
\name{yyhat_view}
\alias{yyhat_view}
\title{Create a prediction vs observation \code{plotly}.}
\usage{
yyhat_view(
  cheem_ls,
  primary_obs = NULL,
  comparison_obs = NULL,
  color = NULL,
  shape = NULL,
  height_px = 480L,
  width_px = 480L
)
}
\arguments{
\item{cheem_ls}{A return from \code{cheem_ls()}, a list of data frames.}

\item{primary_obs}{The rownumber of the primary observation. Its local
attribution becomes the 1d projection basis, and the point it highlighted
as a dashed line.}

\item{comparison_obs}{The rownumber of the comparison observation. Point
is highlighted as a dotted line.}

\item{color}{A vector to map to the point color.
Classification case defaults to predicted class, regression case defaults to
class if passed to cheem_ls(), else residual.}

\item{shape}{A vector to map to the point shape.
Classification case defaults to predicted class, regression case defaults to
class.}

\item{height_px}{The height in pixels of the returned \code{plotly} plot.}

\item{width_px}{The width in pixels of the returned \code{plotly} plot.}
}
\value{
\code{plotly} html widget of the global view, first 2 components of the basis of
the data- and attribution- spaces.
}
\description{
Given an the model from a cheem_ls() list,
Creates a \code{plotly} plot of of the observation predictions by observations.
Typically consumed directly by shiny app.
}
\examples{
library(cheem)

sub <- amesHousing2018_thin[1:200, ]
X <- sub[, 1:10]
Y <- log(sub$SalePrice)
clas <- sub$MS.Zoning

rf_fit <- default_rf(X, Y)
## Long runtime for full datasets:
shap_df <- attr_df_treeshap(rf_fit, X, noisy = FALSE)
this_ls <- cheem_ls(X, Y, class = clas,
                     model = rf_fit,
                     attr_df = shap_df)
yyhat_view(this_ls, 1, 2)
}
\seealso{
Other cheem consumers: 
\code{\link{global_view}()},
\code{\link{radial_cheem_tour}()},
\code{\link{run_app}()}
}
\concept{cheem consumers}
