% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_visualization.r
\name{basis_attr_df}
\alias{basis_attr_df}
\title{Extract and format the 1D local attribution basis}
\usage{
basis_attr_df(attr_df, rownum = 1)
}
\arguments{
\item{attr_df}{A data frame of local explanation attributions,
such as a return from \code{attr_df_treeshap()}.}

\item{rownum}{The rownumber of the primary observation. Defaults to 1.}
}
\value{
A matrix of the 1D basis.
}
\description{
Internal function, Extract and format the 1D local attribution basis from
the provided local explanation's attribution.
}
\examples{
library(cheem)
## Regression:
sub <- amesHousing2018_thin[1:200, ]
X <- sub[, 1:9]
Y <- log(sub$SalePrice)
clas <- sub$ZoneMS

rf_fit  <- default_rf(X, Y)
## Long runtime for full datasets:
shap_df <- attr_df_treeshap(rf_fit, X, noisy = FALSE)
basis_attr_df(shap_df, rownum = 1)
}
