% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{cheem_ls}
\alias{cheem_ls}
\title{Create the local attribution layer data.frame}
\usage{
cheem_ls(
  x,
  y,
  class = NULL,
  model,
  attr_df,
  basis_type = c("pca", "olda"),
  layer_name = class(model)[1],
  verbose = TRUE,
  keep_model = FALSE
)
}
\arguments{
\item{x}{The explanatory variables of the model.}

\item{y}{The target variable of the model.}

\item{class}{The variable to group points by. Originally the \emph{predicted}
class.}

\item{basis_type}{The type of basis used to approximate the data and
attribution space from. Expects "pca" or "olda" (requires \code{clas}).
Defaults to "pca".}

\item{layer_name}{Character layer name, typically the type of local
attribution used.}

\item{verbose}{Logical, if runtime should be printed. Defaults to TRUE.}

\item{noisy}{Logical, if a tone should be played on completion.
Defaults to TRUE.}
}
\value{
A data.frame, for the full local attribution matrix.
}
\description{
Internal function, the local attribution layer data.frame of 1 layer,
consumed
}
\examples{
library(cheem)
## Regression:
x <- sub[, 2:5]
y <- sub$m2.price
clas <- sub$district

.rf_fit <- default_rf(x, y)
.shap_df <- attr_df_treeshap(.rf_fit, x)
out_ls <- cheem_ls(x, y, class = clas,
                   model = .rf_fit,
                   attr_df = .shap_df)
names(out_ls)

## Save for used with shiny app (expects .rds):
if(F) ## Don't accidentally save.
  saveRDS(out_ls, "./my_cheem_ls.rds")
}
