% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{cheem_ls}
\alias{cheem_ls}
\title{Create the local attribution layer data.frame}
\usage{
cheem_ls(
  x,
  y,
  class = NULL,
  model,
  attr_df,
  basis_type = c("pca", "olda"),
  layer_name = class(model)[length(class(model))],
  verbose = TRUE,
  keep_model = FALSE
)
}
\arguments{
\item{x}{The explanatory variables of the model.}

\item{y}{The target variable of the model.}

\item{class}{The variable to group points by. Originally the \emph{predicted}
class.}

\item{model}{A non-linear model, originally a \code{randomForest::randomForest}
model fit, or a return from \code{default_rf()}.}

\item{attr_df}{A data frame of local explanation attributions,
such as a return from \code{attr_df_treeshap()}.}

\item{basis_type}{The type of basis used to approximate the data and
attribution space from. Expects "pca" or "olda" (requires \code{clas}).
Defaults to "pca".}

\item{layer_name}{Character layer name, typically the type of local
attribution used. Defaults to the last class of the model.}

\item{verbose}{Logical, if runtime should be printed. Defaults to TRUE.}

\item{keep_model}{Logical, if the heavy model object should be kept.
Defaults to FALSE.}
}
\value{
A data.frame, for the full local attribution matrix.
}
\description{
Internal function, the local attribution layer data.frame of 1 layer,
consumed
}
\examples{
library(cheem)
library(spinifex)

## Classification:
X    <- penguins[, 1:4]
clas <- penguins$species
Y    <- as.integer(clas)

rf_fit  <- default_rf(X, Y)
## Long runtime for full datasets:
shap_df <- attr_df_treeshap(rf_fit, X, noisy = FALSE)
this_ls <- cheem_ls(X, Y, class = clas,
                     model = rf_fit,
                     attr_df = shap_df)

## Regression:
sub <- amesHousing2018_thin[1:200, ]
X <- sub[, 1:9]
Y <- log(sub$SalePrice)
clas <- sub$ZoneMS

rf_fit  <- default_rf(X, Y)
## Long runtime for full datasets:
shap_df <- attr_df_treeshap(rf_fit, X, noisy = FALSE)
this_ls <- cheem_ls(X, Y, class = clas,
                     model = rf_fit,
                     attr_df = shap_df)

## Save for used with shiny app (expects .rds):
if(FALSE) ## Don't accidentally save.
  saveRDS(this_ls, "./my_cheem_ls.rds")
}
