% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_attribution_ls.r
\name{cheem_ls}
\alias{cheem_ls}
\title{Format the nested local attribution layers}
\usage{
cheem_ls(
  x,
  y,
  xtest = NULL,
  ytest = NULL,
  basis_type = c("pca", "olda"),
  class = NULL,
  loc_attr_nm = "SHAP",
  keep_model = FALSE,
  verbose = TRUE,
  noisy = TRUE
)
}
\arguments{
\item{x}{The explanatory variables of the model.}

\item{y}{The target variable of the model}

\item{xtest}{Optional, Out Of Sample data to find the local attribution of.}

\item{ytest}{Optional, Out Of Sample response to measure xtext with
if provided.}

\item{basis_type}{The type of basis used to approximate the data and
attribution space from. Defaults to "pca".}

\item{class}{The variable to group points by. Originally the \emph{predicted}
class.}

\item{loc_attr_nm}{Name of the layer/attribution. Defaults to "SHAP".}

\item{keep_model}{Whether or not to keep the (sizable) model object.
Defaults to FALSE.}

\item{verbose}{Logical, Whether or not the function should print tictoc time
info. Defaults to TRUE.}

\item{noisy}{Logical, Whether or not the function should play a beepr tone
upon completion. Defaults to TRUE.}
}
\value{
A list of formatted data frames.
}
\description{
Internal function, formats all layers of all data.frames, consumed in
local_attr_ls().
}
\examples{
sub <- DALEX::apartments[1:200, 1:6]
set.seed(303)
.idx_test <- sample(
  1:nrow(sub), size = round(.2 * nrow(sub))) ## 20\% index for testing
X_train <- sub[-.idx_test, 2:5]
Y_train <- sub$m2.price[-.idx_test]
clas <- sub$district[-.idx_test]
X_test  <- sub[.idx_test, 2:5]
Y_test  <- sub$m2.price[.idx_test]

.cheem_ls <- cheem_ls(
  x = X_train, y = Y_train, xtest = X_test, ytest = Y_test,
  basis_type = "pca", class = clas,
  keep_model = FALSE, verbose = TRUE, noisy = TRUE)
names(.cheem_ls)

## Save for used with shiny app (expects .rds):
if(F) ## don't accidentally save a local file.
  saveRDS(.cheem_ls, "./my_cheem_ls.rds")
}
