% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_visualization.r
\name{global_view}
\alias{global_view}
\title{Linked \code{plotly} display, global view of data and attribution space.}
\usage{
global_view(
  cheem_ls,
  primary_obs = NULL,
  comparison_obs = NULL,
  color = NULL,
  shape = NULL,
  height_px = 480L,
  width_px = 1440L,
  as_ggplot = FALSE
)
}
\arguments{
\item{cheem_ls}{A return from \code{cheem_ls()}, a list of data frames.}

\item{primary_obs}{The rownumber of the primary observation. Its local
attribution becomes the 1d projection basis, and the point it highlighted
as a dashed line.}

\item{comparison_obs}{The rownumber of the comparison observation. Point
is highlighted as a dotted line.}

\item{color}{A vector to map to the point color.
Classification case defaults to predicted class, regression case defaults to
class if passed to cheem_ls(), else residual.}

\item{shape}{A vector to map to the point shape.
Classification case defaults to predicted class, regression case defaults to
class.}

\item{height_px}{The height in pixels of the returned \code{plotly} plot.
Defaults to 480.}

\item{width_px}{The width in pixels of the returned \code{plotly} plot.
Defaults to 1440.}

\item{as_ggplot}{Logical, if TRUE returns the plots before being passed to
\code{plotly} functions.}
}
\value{
\code{plotly} html widget of the global view, first 2 components of the basis of
the data- and attribution- spaces.
}
\description{
from a cheem_ls() list, create a linked \code{plotly} of the global data-
and attribution- spaces. Typically consumed directly by shiny app.
}
\examples{
library(cheem)
dat <- amesHousing2018_NorthAmes
X <- dat[, 1:9]
Y <- log(dat$SalePrice)
clas <- dat$SubclassMS

rf_fit <- default_rf(X, Y)
## Long runtime for full datasets:
shap_df <- attr_df_treeshap(rf_fit, X, noisy = FALSE)
this_ls <- cheem_ls(X, Y, class = clas,
                     model = rf_fit,
                     attr_df = shap_df)
global_view(this_ls)
}
\seealso{
Other cheem consumers: 
\code{\link{radial_cheem_tour}()},
\code{\link{run_app}()}
}
\concept{cheem consumers}
