% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{global_view_df_1layer}
\alias{global_view_df_1layer}
\title{Create the plot data.frame for the global linked plotly display.}
\usage{
global_view_df_1layer(x, class = NULL, label = "label")
}
\arguments{
\item{x}{The explanatory variables of the model.}

\item{class}{Optional, (n, 1) vector, a variable to group points by.
This can be the same as or different from \code{y}, the target variable.}

\item{label}{Optionally provide a character label to store reminder
text for the type of model and local explanation used.
Defaults to "label".}
}
\value{
A data.frame, for the global linked \strong{plotly} display.
}
\description{
Internal function consumed in the cheem workflow.
Produces the plot data.frame of 1 layer. consumed downstream in cheem_ls.
}
