% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{global_view_df_1layer}
\alias{global_view_df_1layer}
\title{Create the plot data.frame for the global linked plotly display.}
\usage{
global_view_df_1layer(
  x,
  class = NULL,
  basis_type = c("pca", "olda"),
  layer_name = utils::tail(class(x), 1)
)
}
\arguments{
\item{x}{The explanatory variables of the model.}

\item{class}{The variable to group points by. Originally the \emph{predicted}
class.}

\item{basis_type}{The type of basis used to approximate the data and
attribution space from. Defaults to "pca".}

\item{layer_name}{Character layer name, typically the type of local
attribution used. Defaults to the name of the last class of x.}
}
\value{
A data.frame, for the global linked \strong{plotly} display.
}
\description{
Internal function consumed in the cheem workflow.
Produces the plot data.frame of 1 layer. consumed downstream in cheem_ls.
}
