% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{logistic_tform}
\alias{logistic_tform}
\title{Logistic function to help set alpha opacity}
\usage{
logistic_tform(
  n,
  mid_pt = 1000L,
  k_attenuation = 0.01,
  ceiling = 1L,
  floor = 0.2
)
}
\arguments{
\item{n}{Number of observations to plot.}

\item{mid_pt}{Inflection point that the logistic curve. Defaults to 1000.}

\item{k_attenuation}{The steepness of the transition, larger is more
discrete transition. Quite sensitive and defaults to 0.01.}

\item{ceiling}{The highest number returned. Defaults to 1.}

\item{floor}{The lowest number returned. Defaults to 0.2.}
}
\value{
A scalar numeric, suggested value to set alpha opacity.
}
\description{
Suggests a alpha opacity to plot with as a function of the number of
observation.
}
\examples{
## Suggest an opacity to use in plotting:
(my_alpha <- logistic_tform(nrow(spinifex::penguins)))

## Visualize
x <- 1:2000
plot(x, logistic_tform(x), col='blue')
}
