% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{model_performance_df}
\alias{model_performance_df}
\title{Extract higher level model performance}
\usage{
model_performance_df(model)
}
\arguments{
\item{model}{A model object}

\item{x}{The explanatory data (without response) to extract the local
attributions of.}

\item{y}{The target variable of the model.}

\item{xtest}{Optional, out Of Sample data to find the local attribution of.}

\item{ytest}{Optional, out Of Sample response to measure x with
if provided.}
}
\value{
A dataframe of the local attributions.
}
\description{
Internal function, used downstream in cheem_ls.
}
\examples{
library(cheem)
sub <- DALEX::apartments[1:200, 1:5]
x <- sub[, 2:5]
y <- sub$m2.price

.rf_fit <- default_rf(x, y)
model_performance_df(.rf_fit, x)
}
