% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{model_performance_df}
\alias{model_performance_df}
\title{Extract higher level model performance statistics}
\usage{
model_performance_df(model, x = NULL, y = NULL)
}
\arguments{
\item{model}{A non-linear model, originally a \code{randomForest::randomForest}
model fit, or a return from \code{default_rf()}.}

\item{x}{Data to predict, required by ranger models.}

\item{y}{Observed response, required by ranger models.}
}
\value{
A data.frame of model performance statistics.
}
\description{
Internal function, used downstream in cheem_ls.
}
\examples{
library(cheem)

## Regression setup:
dat  <- amesHousing2018_NorthAmes
X    <- dat[, 1:9]
Y    <- log(dat$SalePrice)
clas <- dat$SubclassMS

## Model and performance:
rf_fit <- default_rf(X, Y)
cheem:::model_performance_df(rf_fit)
}
