% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_cheem_lists.r
\name{model_performance_df}
\alias{model_performance_df}
\title{Extract higher level model performance}
\usage{
model_performance_df(model)
}
\arguments{
\item{model}{A non-linear model, originally a \code{randomForest::randomForest}
model fit, or a return from \code{default_rf()}.}
}
\value{
A dataframe of the local attributions.
}
\description{
Internal function, used downstream in cheem_ls.
}
\examples{
library(cheem)
## Regression:
sub <- amesHousing2018_thin[1:200, ]
X <- sub[, 1:9]
Y <- log(sub$SalePrice)
clas <- sub$ZoneMS

rf_fit <- default_rf(X, Y)
model_performance_df(rf_fit, X)
}
