% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_util.r
\name{problem_type}
\alias{problem_type}
\title{The type of model for a given Y variable.}
\usage{
problem_type(y)
}
\arguments{
\item{y}{Response variable to be modeled}
}
\value{
Character in c("classification", "regression")
}
\description{
Whether the Y is a "classification", "regression" or ill-defined problem.
Returns a character: "classification", "regression", or an error for strange
classes. Minor redundancy with is_discrete, though explicit, and does check
for strange classes.
}
\examples{
problem_type(mtcars$mpg)
problem_type(mtcars$cyl) ## Numeric column, labeled as discrete, because less than 9 unique values
problem_type(letters)
}
