% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_visualization.r
\name{proto_basis1d_distribution}
\alias{proto_basis1d_distribution}
\title{Adds the distribution of the row local attributions to a ggtour}
\usage{
proto_basis1d_distribution(
  attr_df,
  group_by = as.factor(FALSE),
  position = c("top1d", "floor1d", "top2d", "floor2d", "off"),
  shape = c(142, 124),
  do_add_pcp_segments = TRUE,
  primary_obs = 1,
  comparison_obs = 2,
  inc_vars = NULL,
  row_index = NULL
)
}
\arguments{
\item{group_by}{Vector to group densities by. Originally \emph{predicted} class.}

\item{position}{The position for the basis, one of: c("top1d", "floor1d",
"top2d", "floor2d", "off").
Defaults to "top1d"; basis above the density curves.}

\item{shape}{Number specifying the shape of the basis distribution points.
Typically 142 or 124 indicating '|' for \code{plotly} and \code{gganimate} respectively.
Defaults to 142, '|' for \code{plotly}.}

\item{do_add_pcp_segments}{Logical, whether or not to add to add faint
parallel coordinate lines on the 1D basis. Defaults to TRUE.}

\item{primary_obs}{The rownumber of the primary observation. Its local
attribution becomes the 1d projection basis, and the point it highlighted
as a dashed line.}

\item{comparison_obs}{The rownumber of the comparison observation. Point
is highlighted as a dotted line.}

\item{cheem_ls}{A return from \code{cheem_ls()}, a list of data frames.}
}
\description{
Adds the distribution of orthonormalized row values of
the specified local explanation \code{attr_df}. Does not draw the basis itself;
use in conjunction with \code{proto_basis1d()}.
}
\examples{
library(cheem)
library(spinifex)
## Classification:
X    <- flea[, 1:6]
clas <- flea$species
Y    <- as.integer(clas)

rf_fit  <- default_rf(X, Y)
shap_df <- attr_df_treeshap(rf_fit, X) ## Long runtime!
this_ls <- cheem_ls(X, Y, class = clas,
                    model = rf_fit,
                    attr_df = shap_df)

bas <- basis_local_attribution(shap_df, rownum = 1)
mv <- manip_var_of(bas)
mt_path <- manual_tour(bas, mv)

X_scl <- spinifex::scale_sd(X)
ggt <- ggtour(mt_path, X_scl, angle = .3) +
  proto_density(aes_args = list(color = clas, fill = clas)) +
  proto_basis1d() +
  proto_basis1d_distribution(shap_df, group_by = clas)
\dontrun{
animate_plotly(ggt)
}

library(cheem)
library(spinifex)
## Regression:
sub <- amesHousing2018_thin[1:200, ]
X <- sub[, 1:10]
Y <- log(sub$SalePrice)
clas <- sub$MS.Zoning

rf_fit  <- default_rf(X, Y)
shap_df <- attr_df_treeshap(rf_fit, X) ## Long runtime!
this_ls <- cheem_ls(X, Y, class = clas,
                    model = rf_fit,
                    attr_df = shap_df)

bas <- basis_local_attribution(shap_df, rownum = 1)
mv  <- manip_var_of(bas)
ggt <- radial_cheem_ggtour(
  this_ls, basis = bas, manip_var = mv, angle = .3)
\dontrun{
animate_plotly(ggt)
}
}
\concept{ggtour proto}
