% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_visualization.r
\name{proto_basis1d_distribution}
\alias{proto_basis1d_distribution}
\title{Adds the distribution of the row local attributions to a ggtour}
\usage{
proto_basis1d_distribution(
  attr_df,
  group_by = as.factor(FALSE),
  position = c("top1d", "floor1d", "bottom1d", "off"),
  shape = c(142, 124),
  do_add_pcp_segments = TRUE,
  primary_obs = 1,
  comparison_obs = 2,
  inc_vars = NULL,
  row_index = NULL
)
}
\arguments{
\item{attr_df}{An data frame, the attributions of a local explanation,
such as a return from \code{attr_df_treeshap()}.}

\item{group_by}{Vector to group densities by. Originally \emph{predicted} class.}

\item{position}{The position for the basis, one of: c("top1d", "floor1d",
"bottom1d", "off").
Defaults to "top1d"; basis above the density curves.}

\item{shape}{Number specifying the shape of the basis distribution points.
Typically 142 or 124 indicating '|' for \code{plotly} and \code{gganimate} respectively.
Defaults to 142, '|' for \code{plotly}.}

\item{do_add_pcp_segments}{Logical, whether or not to add to add faint
parallel coordinate lines on the 1D basis. Defaults to TRUE.}

\item{primary_obs}{The rownumber of the primary observation. Its local
attribution becomes the 1d projection basis, and the point it highlighted
as a dashed line.}

\item{comparison_obs}{The rownumber of the comparison observation. Point
is highlighted as a dotted line.}

\item{inc_vars}{A character vector, the names of the variables to keep.
Defaults to NULL, all variables kept.}

\item{row_index}{A numeric or logical vector, the index of the rows to keep.
Defaults to NULL, all rows kept.}
}
\description{
A {spinifex} proto_*-like function, that adds the distribution of
orthonormalized row values of the specified local explanation \code{attr_df}.
Does not draw the basis bars; use in conjunction with \code{proto_basis1d()}.
}
\examples{
library(cheem)
library(spinifex)

sub <- amesHousing2018_thin[1:200, ]
X <- sub[, 1:9]
Y <- log(sub$SalePrice)
clas <- sub$ZoneMS

rf_fit <- default_rf(X, Y)
## Long runtime for full datasets:
shap_df <- attr_df_treeshap(rf_fit, X, noisy = FALSE)

bas_p <- basis_attr_df(shap_df, rownum = 1)
bas_c <- basis_attr_df(shap_df, rownum = 2)
diff <- abs(bas_p - bas_c)
mv <- which(diff == max(diff))
mt_path <- manual_tour(bas_p, mv)
fixed_pred <- stats::predict(rf_fit)

ggt <- ggtour(mt_path, scale_sd(X), angle = .3) +
  append_fixed_y(fixed_y = scale_sd(fixed_pred)) +
  proto_point(list(color = clas, shape = clas)) +
  proto_basis1d_distribution(
    attr_df = shap_df, group_by = clas, position = "top1d",
    primary_obs = 1, comparison_obs = 2) +
  proto_basis1d(position = "bottom1d") +
  proto_origin()
\dontrun{
animate_plotly(ggt)
}
}
\concept{ggtour proto}
