% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_run_app.r
\name{run_app}
\alias{run_app}
\title{Runs a shiny app demonstrating manual tours}
\usage{
run_app(app_nm = "cheem", ...)
}
\arguments{
\item{app_nm}{Name of the shiny app to run. Expects "cheem_initial".}

\item{...}{Other arguments passed into \code{shiny::runApp()}.
Such as display.mode = "showcase".}
}
\value{
Runs a locally hosted shiny app.
}
\description{
Runs a local shiny app that demonstrates manual tour and comparable
traditional techniques for static projections of multivariate data sets.
}
\examples{
## Only run this example in interactive R sessions
if(interactive()){
  ## Runs the app
  run_app("cheem")
  
  ## Run with app code displayed
  run_app(app_nm = "cheem", display.mode = "showcase")
}
}
\seealso{
Other cheem consumers: 
\code{\link{global_view_legwork}()},
\code{\link{global_view}()},
\code{\link{radial_cheem_tour}()}
}
\concept{cheem consumers}
